% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.helpersAbsRisk.R
\name{makeCifNpDf}
\alias{makeCifNpDf}
\title{Prepare Non-Parametric CIF Estimates}
\usage{
makeCifNpDf(transprob.out)
}
\arguments{
\item{transprob.out}{Output from `etm::etm` or similar, representing transition probabilities.}
}
\value{
A `data.frame` with CIF estimates and confidence intervals.
}
\description{
Converts the output from a non-parametric CIF estimation into a tidy format for plotting.
}
