% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.helpersAbsRisk.R
\name{makeCifSpDf}
\alias{makeCifSpDf}
\title{Prepare Semi-Parametric CIF Estimates}
\usage{
makeCifSpDf(transprob.out)
}
\arguments{
\item{transprob.out}{A list with time-varying probabilities for each state, typically from `mstate` or similar.}
}
\value{
A `data.frame` suitable for plotting.
}
\description{
Converts semi-parametric transition probability output into a tidy format for plotting.
}
