% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.helpersAbsRisk.R
\name{makeCumhazDf}
\alias{makeCumhazDf}
\title{Prepare Non-Parametric Cumulative Hazard Data}
\usage{
makeCumhazDf(cumhaz.out)
}
\arguments{
\item{cumhaz.out}{A list of two data frames for the cumulative hazard of each transition, typically output from `prodlim::prodlim`.}
}
\value{
A `data.frame` with hazard estimates and confidence bounds.
}
\description{
Converts cumulative hazard estimates (e.g., from Aalen-Johansen estimator) into a tidy format with confidence intervals.
}
