% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.helpersAbsRisk.R
\name{plot_AbsRiskCox2Events}
\alias{plot_AbsRiskCox2Events}
\title{Plot Cause-Specific Cumulative Hazards (Semi-Parametric)}
\usage{
plot_AbsRiskCox2Events(cox_e1, cox_e2, cols = default_cols)
}
\arguments{
\item{cox_e1}{Output from a Cox model (e.g., `survfit`) for event 1.}

\item{cox_e2}{Output from a Cox model for event 2.}

\item{cols}{Vector of colors for each transition line. Default is `c("black", "darkgrey")`.}
}
\value{
A `ggplot` object visualizing the cumulative hazard functions.
}
\description{
Plots the cumulative hazard estimates from two Cox model outputs using ggplot2, including 95% confidence intervals.
}
