% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.helpersAbsRisk.R
\name{plot_AbsRiskSpaj2Events}
\alias{plot_AbsRiskSpaj2Events}
\title{Plot Semi-Parametric Transition Probabilities (Aalen-Johansen)}
\usage{
plot_AbsRiskSpaj2Events(transprob, cols = default_cols)
}
\arguments{
\item{transprob}{Output from a semi-parametric transition probability estimator.}

\item{cols}{Vector of colors for each transition. Default is `c("black", "darkgrey")`.}
}
\value{
A `ggplot` object.
}
\description{
Plots semi-parametric transition probability estimates over time using a Cox-Aalen-Johansen estimator.
}
