% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.compareROC.R
\name{plot_compareROC}
\alias{plot_compareROC}
\title{Plot a comparison of weighted ROC Curve using ggplot2}
\usage{
plot_compareROC(x, y, label, names, colors, ...)
}
\arguments{
\item{x}{A nested list, with for each comparison a list object returned by `weightedRoc()` containing TPR, FPR, AUC, etc.}

\item{y}{A nested list, with for each comparison a list object returned by `aucWithCI()` containing AUC and its confidence interval.}

\item{label}{Group labels to be used in the plot}

\item{names}{Group names}

\item{colors}{Line colors for the different groups}

\item{...}{Additional arguments passed to `ggplot2::theme()` for plot customization.}
}
\value{
A `ggplot` object showing the ROC curve with AUC and confidence interval annotated.
}
\description{
Generates a ggplot of the two or more weighted ROC curves at a specified evaluation time `t_eval`,
using the output from `weightedRoc()` and `aucWithCI()`.
}
\examples{
\dontrun{
#' roc_obj1 <- weightedRoc(data, t_eval = 5, marker = "biomarker")
auc_ci1  <- aucWithCI(data, t_eval = 5, marker = "biomarker")
roc_obj2 <- weightedRoc(data, t_eval = 5, marker = "biomarker")
auc_ci2  <- aucWithCI(data, t_eval = 5, marker = "biomarker")
roc_obj_list <- list(roc_obj1,roc_obj2)
auc_ci_list <- list(auc_ci1,auc_ci2)
plot_compareROC(roc_obj_list, auc_ci_list, t_eval = 5,
               names = c(cohort1, cohort2), colors = c("purple", "pink"))
}
}
