% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.ROC.R
\name{plot_ROC}
\alias{plot_ROC}
\title{Plot Weighted ROC Curve using ggplot2}
\usage{
plot_ROC(x, y, t_eval, ...)
}
\arguments{
\item{x}{A list object returned by `weightedRoc()` containing TPR, FPR, AUC, etc.}

\item{y}{A list object returned by `aucWithCI()` containing AUC and its confidence interval.}

\item{t_eval}{Numeric value for the time point at which AUC is evaluated.}

\item{...}{Additional arguments passed to `ggplot2::theme()` for plot customization.}
}
\value{
A `ggplot` object showing the ROC curve with AUC and confidence interval annotated.
}
\description{
Generates a ggplot of the weighted ROC curve at a specified evaluation time `t_eval`,
using the output from `weightedRoc()` and `aucWithCI()`.
}
\examples{
\dontrun{
roc_obj <- weightedRoc(data, t_eval = 5, marker = "biomarker")
auc_ci  <- aucWithCI(data, t_eval = 5, marker = "biomarker")
plot_ROC(roc_obj, auc_ci, 5)
}

}
