% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.timeROC.R
\name{plot_timeROC}
\alias{plot_timeROC}
\title{Plot ROC Curve from timeROC Output Using ggplot2}
\usage{
plot_timeROC(timeROC_obj, ci_df, definition = 2, time_index = 2, ...)
}
\arguments{
\item{timeROC_obj}{Output object from the `timeROC()` function.}

\item{ci_df}{A `data.frame` from `confint(timeROC_obj)` representing AUC confidence intervals.}

\item{definition}{Integer; either 1 or 2.
\itemize{
  \item `1`: Competing events do not contribute to AUC (Definition 1).
  \item `2`: Competing risks treated as controls (Definition 2).
}}

\item{time_index}{Integer index of the `times` value in the `timeROC` object to use for plotting (default = 2).}

\item{...}{Additional arguments passed to `ggplot2::theme()` for plot customization.}
}
\value{
A `ggplot` object of the ROC curve with annotated AUC and 95% CI.
}
\description{
This helper function creates a ROC plot from the output of the `timeROC` function,
using either Definition 1 or Definition 2 for handling competing risks.
Note that it only works for one time_index = 2, with times = c() a vector of length 2, and the first element = 0.
}
\examples{
\dontrun{
test <- timeROC(T = cohort$time, delta = cohort$status, marker = cohort$X,
                cause = 1, times = c(0, 1826.25), weighting = "marginal")
result.ci <- confint(test, level = 0.95, n.sim = 1000) \%>\% as.data.frame()
plot_timeROC(test, result.ci, definition = 2)
}

}
